roo-standalone/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.2.0
alembic>=1.12.0

# Slack
slack-sdk>=3.21.0

# LLM Providers
openai>=1.3.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Utilities
python-frontmatter>=1.0.0
httpx>=0.25.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
